add authentication login api

+++ server/routers/deleted.js
@@ -53,40 +53,46 @@ module.exports = function (bot, time, changeTime, db) {
     router.delete(
         "/purchased_products/:productId/:userId/:nameProduct",
         (req, res) => {
-            const productId = req.params.productId;
-            const nameProduct = req.params.nameProduct;
-            const userId = req.params.userId;
-            const sql = "DELETE FROM purchased_products WHERE id = ?";
-            const infoUser = `SELECT * FROM users WHERE id = ?`;
-            const updateDateQuery = `UPDATE users SET created_at = ? WHERE id = ?`;
+            if (req.session && req.session.loggedIn) {
 
-            db.run(updateDateQuery, [time(), userId]);
-            db.run(sql, [productId], (err) => {
-                if (err) {
-                    console.error(err.message);
-                    return res.json({ error: StatusCodes.INTERNAL_SERVER_ERROR, message: "Failed to add received price" });
-                }
-                res.json({ success: true });
-                db.get(infoUser, [userId], (err, row) => {
+                const productId = req.params.productId;
+                const nameProduct = req.params.nameProduct;
+                const userId = req.params.userId;
+                const sql = "DELETE FROM purchased_products WHERE id = ?";
+                const infoUser = `SELECT * FROM users WHERE id = ?`;
+                const updateDateQuery = `UPDATE users SET created_at = ? WHERE id = ?`;
+
+                db.run(updateDateQuery, [time(), userId]);
+                db.run(sql, [productId], (err) => {
                     if (err) {
                         console.error(err.message);
                         return res.json({ error: StatusCodes.INTERNAL_SERVER_ERROR, message: "Failed to add received price" });
                     }
-                    bot.telegram.sendDocument(
-                        process.env.BOT_ADMIN,
-                        { source: "./db/customer.db" },
-                        {
-                            caption: `<b><u>حذف کالا :</u></b>
+                    res.json({ success: true });
+                    db.get(infoUser, [userId], (err, row) => {
+                        if (err) {
+                            console.error(err.message);
+                            return res.json({ error: StatusCodes.INTERNAL_SERVER_ERROR, message: "Failed to add received price" });
+                        }
+                        bot.telegram.sendDocument(
+                            process.env.BOT_ADMIN,
+                            { source: "./db/customer.db" },
+                            {
+                                caption: `<b><u>حذف کالا :</u></b>
   
   در : <code>${changeTime(time())}</code>
   نام مشتری : <b>${row.name} ${row.family}</b>
   نام محصول : <b>${nameProduct}</b>
   .`,
-                            parse_mode: "HTML",
-                        }
-                    );
+                                parse_mode: "HTML",
+                            }
+                        );
+                    });
                 });
-            });
+            } else {
+                // User is not logged in or session is not saved
+                res.status(StatusCodes.UNAUTHORIZED).json({ error: "Unauthorized" });
+            }
         }
     );
 
@@ -98,40 +104,46 @@ module.exports = function (bot, time, changeTime, db) {
     router.delete(
         "/recived_price/:productId/:userId/:nameProduct",
         (req, res) => {
-            const productId = req.params.productId;
-            const nameProduct = req.params.nameProduct;
-            const userId = req.params.userId;
-            const sql = "DELETE FROM recived_price WHERE id = ?";
-            const infoUser = `SELECT * FROM users WHERE id = ?`;
-            const updateDateQuery = `UPDATE users SET created_at = ? WHERE id = ?`;
+            if (req.session && req.session.loggedIn) {
 
-            db.run(updateDateQuery, [time(), userId]);
-            db.run(sql, [productId], (err) => {
-                if (err) {
-                    console.error(err.message);
-                    return res.json({ error: StatusCodes.INTERNAL_SERVER_ERROR, message: "Failed to add received price" });
-                }
-                res.json({ success: true });
-                db.get(infoUser, [userId], (err, row) => {
+                const productId = req.params.productId;
+                const nameProduct = req.params.nameProduct;
+                const userId = req.params.userId;
+                const sql = "DELETE FROM recived_price WHERE id = ?";
+                const infoUser = `SELECT * FROM users WHERE id = ?`;
+                const updateDateQuery = `UPDATE users SET created_at = ? WHERE id = ?`;
+
+                db.run(updateDateQuery, [time(), userId]);
+                db.run(sql, [productId], (err) => {
                     if (err) {
                         console.error(err.message);
                         return res.json({ error: StatusCodes.INTERNAL_SERVER_ERROR, message: "Failed to add received price" });
                     }
-                    bot.telegram.sendDocument(
-                        process.env.BOT_ADMIN,
-                        { source: "./db/customer.db" },
-                        {
-                            caption: `<b><u>حذف کالا :</u></b>
+                    res.json({ success: true });
+                    db.get(infoUser, [userId], (err, row) => {
+                        if (err) {
+                            console.error(err.message);
+                            return res.json({ error: StatusCodes.INTERNAL_SERVER_ERROR, message: "Failed to add received price" });
+                        }
+                        bot.telegram.sendDocument(
+                            process.env.BOT_ADMIN,
+                            { source: "./db/customer.db" },
+                            {
+                                caption: `<b><u>حذف کالا :</u></b>
   
   در : <code>${changeTime(time())}</code>
   نام مشتری : <b>${row.name} ${row.family}</b>
   نام محصول : <b>${nameProduct}</b>
   .`,
-                            parse_mode: "HTML",
-                        }
-                    );
+                                parse_mode: "HTML",
+                            }
+                        );
+                    });
                 });
-            });
+            } else {
+                // User is not logged in or session is not saved
+                res.status(StatusCodes.UNAUTHORIZED).json({ error: "Unauthorized" });
+            }
         }
     );
 
@@ -141,34 +153,40 @@ module.exports = function (bot, time, changeTime, db) {
  * @param {Object} res - The response object.
  */
     router.delete("/users/:id/:name/:family", (req, res) => {
-        const Id = req.params.id;
-        const name = req.params.name;
-        const family = req.params.family;
-        const sql = "DELETE FROM users WHERE id = ?";
-        const sql1 = "DELETE FROM purchased_products WHERE user_id = ?";
-        const sql2 = "DELETE FROM recived_price WHERE user_id = ?";
+        if (req.session && req.session.loggedIn) {
 
-        db.run(sql, [Id], (err) => {
-            if (err) {
-                console.error(err.message);
-                return res.json({ error: StatusCodes.INTERNAL_SERVER_ERROR, message: "Failed to add received price" });
-            }
-            res.json({ success: true });
-            bot.telegram.sendDocument(
-                process.env.BOT_ADMIN,
-                { source: "./db/customer.db" },
-                {
-                    caption: `<b><u>حذف مشتری :</u></b>
+            const Id = req.params.id;
+            const name = req.params.name;
+            const family = req.params.family;
+            const sql = "DELETE FROM users WHERE id = ?";
+            const sql1 = "DELETE FROM purchased_products WHERE user_id = ?";
+            const sql2 = "DELETE FROM recived_price WHERE user_id = ?";
+
+            db.run(sql, [Id], (err) => {
+                if (err) {
+                    console.error(err.message);
+                    return res.json({ error: StatusCodes.INTERNAL_SERVER_ERROR, message: "Failed to add received price" });
+                }
+                res.json({ success: true });
+                bot.telegram.sendDocument(
+                    process.env.BOT_ADMIN,
+                    { source: "./db/customer.db" },
+                    {
+                        caption: `<b><u>حذف مشتری :</u></b>
   
   در : <code>${changeTime(time())}</code>
   نام مشتری : <b>${name} ${family}</b>
   .`,
-                    parse_mode: "HTML",
-                }
-            );
-        });
-        db.run(sql1, [Id]);
-        db.run(sql2, [Id]);
+                        parse_mode: "HTML",
+                    }
+                );
+            });
+            db.run(sql1, [Id]);
+            db.run(sql2, [Id]);
+        } else {
+            // User is not logged in or session is not saved
+            res.status(StatusCodes.UNAUTHORIZED).json({ error: "Unauthorized" });
+        }
     });
 
     return router
